Fixed decals in orthographic mode; fixed index out of bounds for FloatImage; fixed MeshBuilder/MeshSplitter - flags were incorrectly defined & faces were flipped on cut surfaces; made Systems cleaner by not extending OnUpdate; fixed RenderMode in thumbnails - depth seems to be broken now 🤔; fixed margin-type-discrepancy in a few tests

+++ test/src/me/anno/tests/physics/shapes/SDFCubeTest.kt
@@ -19,7 +19,7 @@ class SDFCubeTest {
     private fun createBaseline(marginI: Float): BoxShape {
         val baseline = BoxCollider().apply { margin = marginI }
             .createBulletShape(Vector3d(1.0))
-        assertEquals(marginI, baseline.margin)
+        assertEquals(marginI, baseline.margin.toFloat())
         return baseline
     }
